+++ traffic_control/clients/python/trafficops/tosession.py
@@ -352,17 +352,6 @@ class TOSession(RestApiSession):
 		:raises: Union[LoginError, OperationError]
 		"""
 
-	@api_request('get', 'cachegroups/{cache_group_id:d}/unassigned_parameters', ('2.0',))
-	def get_cachegroup_unassigned_parameters(self, cache_group_id=None):
-		"""
-		Get a cache groups unassigned parameters
-		:ref:`to-api-cachegroups-id-unassigned_parameters`
-		:param cache_group_id: The cache group Id
-		:type cache_group_id: int
-		:rtype: Tuple[Dict[str, Any], requests.Response]
-		:raises: Union[LoginError, OperationError]
-		"""
-
 	@api_request('get', 'cachegroupparameters', ('2.0',))
 	def get_all_cachegroup_parameters(self):
 		"""